pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miura-forge-engine"
version = "2.0.0"
description = "Motor avanzado de automatización de contenido para YouTube Shorts"
readme = "Docs/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Miura Forge Team"}
]
keywords = ["youtube", "automation", "ai", "content-creation", "shorts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "google-genai>=0.3.0",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "gspread>=6.0.0",
    "oauth2client>=4.1.3",
    "replicate>=0.25.0",
    "openai>=1.12.0",
    "playwright>=1.41.0",
    "playwright-stealth>=1.0.6",
    "tavily-python>=0.3.0",
    "moviepy>=1.0.3",
    "movielite>=0.2.0",
    "faster-whisper>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "Pillow>=10.0.0",
    "rich>=13.7.0",
    "alive-progress>=3.1.0",
    "tenacity>=8.2.0",
    "pydantic>=2.6.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyarrow>=15.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
miura-forge = "main:main"
miura-oracle = "tools.weekly_oracle:run_weekly_oracle"

[project.urls]
Homepage = "https://github.com/miura-forge/engine"
Documentation = "https://docs.miura-forge.ai"
Repository = "https://github.com/miura-forge/engine"

[tool.setuptools]
packages = [
    "core",
    "tools",
    "motion_forge",
    "youtube_publisher",
    "deployer",
    "auditoria",
    "llm",
    "merch",
    "emisario",
]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.json", "*.md", "*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | venv
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["core", "tools", "motion_forge"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
